+++ CHANGELOG.md
@@ -23,6 +23,8 @@ response.download(csvData, 'example.csv');
 
 #### server module corrected mime type of json response
 
+#### gracenode core module loader improved
+
 ### Depricated
 
 None